Refactor: split into separate domains
