+++ app/knowledge_management/repository.py
@@ -4,16 +4,11 @@ from bson.datetime_ms import DatetimeMS
 from bson.objectid import ObjectId
 from pymongo.asynchronous.database import AsyncCollection, AsyncDatabase
 from pymongo.errors import DuplicateKeyError
-from sqlalchemy import select
 
-from app.knowledge.models import (
+from app.knowledge_management.models import (
     KnowledgeGroup,
     KnowledgeGroupAlreadyExistsError,
-    KnowledgeResult,
-    KnowledgeSearchResults,
-    KnowledgeSnapshot,
     KnowledgeSource,
-    KnowledgeVector,
 )
 
 
@@ -29,30 +24,6 @@ class AbstractKnowledgeGroupRepository(ABC):
     @abstractmethod
     async def list_all(self) -> list[KnowledgeGroup]:
         """List all knowledge groups with their sources loaded"""
-
-
-class AbstractKnowledgeSnapshotRepository(ABC):
-    @abstractmethod
-    async def save(self, snapshot) -> None:
-        """Save a knowledge snapshot"""
-
-    @abstractmethod
-    async def get_by_id(self, snapshot_id: str):
-        """Get a knowledge snapshot by its ID"""
-
-    @abstractmethod
-    async def list_all(self) -> list[KnowledgeSnapshot]:
-        """List all knowledge snapshots"""
-
-
-class AbstractKnowledgeVectorRepository(ABC):
-    @abstractmethod
-    async def add(self, knowledge_vector: KnowledgeVector) -> None:
-        """Add a knowledge vector entry"""
-
-    @abstractmethod
-    async def query(self, embedding: list[float], top_k: int) -> KnowledgeSearchResults:
-        """Query for the top_k most similar knowledge vectors"""
 
 
 class MongoKnowledgeGroupRepository(AbstractKnowledgeGroupRepository):
@@ -166,109 +137,3 @@ class MongoKnowledgeGroupRepository(AbstractKnowledgeGroupRepository):
             groups.append(group)
 
         return groups
-
-
-class MongoKnowledgeSnapshotRepository(AbstractKnowledgeSnapshotRepository):
-    def __init__(self, db: AsyncDatabase):
-        self.db: AsyncDatabase = db
-        self.knowledge_snapshots: AsyncCollection = self.db.get_collection("knowledgeSnapshots")
-
-    async def save(self, snapshot: KnowledgeSnapshot) -> None:
-        """Save a knowledge snapshot"""
-        snapshot_data = {
-            "snapshotId": snapshot.snapshot_id,
-            "groupId": snapshot.group_id,
-            "version": snapshot.version,
-            "createdAt": DatetimeMS(snapshot.created_at),
-            "sources": [source.__dict__ for source in snapshot.sources]
-        }
-
-        await self.knowledge_snapshots.insert_one(snapshot_data)
-
-    async def get_by_id(self, snapshot_id: str) -> KnowledgeSnapshot | None:
-        """Get a knowledge snapshot by its ID"""
-        doc = await self.knowledge_snapshots.find_one({"snapshotId": snapshot_id})
-        if not doc:
-            return None
-
-        return KnowledgeSnapshot(
-            group_id=doc["groupId"],
-            version=doc["version"],
-            created_at=doc["createdAt"],
-            sources=doc["sources"]
-        )
-
-    async def list_snapshots_by_group(self, group_id: str) -> list[KnowledgeSnapshot]:
-        """List all knowledge snapshots for a specific group"""
-        cursor = self.knowledge_snapshots.find({"groupId": group_id})
-        snapshots = []
-
-        async for doc in cursor:
-            snapshot = KnowledgeSnapshot(
-                group_id=doc["groupId"],
-                version=doc["version"],
-                created_at=doc["createdAt"],
-                sources=doc["sources"]
-            )
-            snapshots.append(snapshot)
-
-        return snapshots
-
-
-class PostgresKnowledgeVectorRepository(AbstractKnowledgeVectorRepository):
-    """PostgreSQL implementation of KnowledgeVectorRepository using pgvector."""
-
-    def __init__(self, session):
-        """
-        Initialize with SQLAlchemy async session.
-
-        Args:
-            session: Async SQLAlchemy session
-        """
-        self.session = session
-
-    async def add(self, knowledge_vector: KnowledgeVector) -> None:
-        """Add a knowledge vector entry to PostgreSQL."""
-        self.session.add(knowledge_vector)
-        await self.session.commit()
-
-    async def add_batch(self, vectors: list[KnowledgeVector]) -> None:
-        """Add multiple knowledge vector entries to PostgreSQL in batch."""
-        self.session.add_all(vectors)
-
-        await self.session.commit()
-
-    async def query(self, embedding: list[float], top_k: int) -> KnowledgeSearchResults:
-        """Query for the top_k most similar knowledge vectors using cosine similarity."""
-        # Use cosine distance for similarity search (lower distance = higher similarity)
-        stmt = (
-            select(
-                KnowledgeVector.id,
-                KnowledgeVector.content,
-                KnowledgeVector.embedding,
-                KnowledgeVector.created_at,
-                KnowledgeVector.embedding.cosine_distance(embedding).label("distance")
-            )
-            .order_by(KnowledgeVector.embedding.cosine_distance(embedding))
-            .limit(top_k)
-        )
-
-        result = await self.session.execute(stmt)
-        rows = result.fetchall()
-
-        # Convert to domain objects
-        vector_results = [
-            KnowledgeResult(
-                content=row.content,
-                similarity_score=1.0 - float(row.distance),  # Convert distance to similarity
-                created_at=row.created_at,
-                embedding=row.embedding  # Include embedding in results
-            )
-            for row in rows
-        ]
-
-        return KnowledgeSearchResults(
-            query_embedding=embedding,
-            results=vector_results
-        )
-
