+++ app/knowledge_management/router.py
@@ -7,24 +7,27 @@ from app.common.bedrock import BedrockEmbeddingService, get_bedrock_client
 from app.common.mongo import get_db
 from app.common.postgres import get_async_session_factory
 from app.config import config
-from app.knowledge.models import (
+from app.ingestion.repository import (
+    AbstractKnowledgeVectorRepository,
+    PostgresKnowledgeVectorRepository,
+)
+from app.ingestion.service import IngestionService
+from app.knowledge_management.models import (
     KnowledgeGroup,
     KnowledgeGroupNotFoundError,
     KnowledgeSource,
 )
-from app.knowledge.repository import (
+from app.knowledge_management.repository import (
     AbstractKnowledgeGroupRepository,
-    AbstractKnowledgeVectorRepository,
     MongoKnowledgeGroupRepository,
-    PostgresKnowledgeVectorRepository,
 )
-from app.knowledge.request_schemas import (
+from app.knowledge_management.request_schemas import (
     CreateKnowledgeGroupRequest,
     KnowledgeGroupResponse,
 )
-from app.knowledge.service import KnowledgeService
+from app.knowledge_management.service import KnowledgeManagementService
 
-router = APIRouter(tags=["knowledge"])
+router = APIRouter(tags=["knowledge-management"])
 
 
 def get_knowledge_repository(db: AsyncDatabase = Depends(get_db)) -> MongoKnowledgeGroupRepository:
@@ -44,13 +47,18 @@ async def get_bedrock_embedding_service():
     return BedrockEmbeddingService(get_bedrock_client(), config.bedrock_embedding_config)
 
 
-def get_knowledge_service(group_repo: AbstractKnowledgeGroupRepository = Depends(get_knowledge_repository), vector_repo: AbstractKnowledgeVectorRepository = Depends(get_knowledge_vector_repository), embedding_service: BedrockEmbeddingService = Depends(get_bedrock_embedding_service)) -> KnowledgeService:
-    """Dependency injection for KnowledgeService."""
-    return KnowledgeService(group_repo, vector_repo, embedding_service)
+def get_knowledge_management_service(group_repo: AbstractKnowledgeGroupRepository = Depends(get_knowledge_repository)) -> KnowledgeManagementService:
+    """Dependency injection for KnowledgeManagementService."""
+    return KnowledgeManagementService(group_repo)
+
+
+def get_ingestion_service(vector_repo: AbstractKnowledgeVectorRepository = Depends(get_knowledge_vector_repository), embedding_service: BedrockEmbeddingService = Depends(get_bedrock_embedding_service)) -> IngestionService:
+    """Dependency injection for IngestionService."""
+    return IngestionService(vector_repo, embedding_service)
 
 
 @router.get("/knowledge/groups", status_code=status.HTTP_200_OK, response_model=list[KnowledgeGroupResponse])
-async def list_groups(service: KnowledgeService = Depends(get_knowledge_service)):
+async def list_groups(service: KnowledgeManagementService = Depends(get_knowledge_management_service)):
     """
     List all knowledge groups.
 
@@ -78,7 +86,7 @@ async def list_groups(service: KnowledgeService = Depends(get_knowledge_service)
 
 
 @router.post("/knowledge/groups", status_code=status.HTTP_201_CREATED, response_model=KnowledgeGroupResponse)
-async def create_group(group: CreateKnowledgeGroupRequest, service: KnowledgeService = Depends(get_knowledge_service)):
+async def create_group(group: CreateKnowledgeGroupRequest, service: KnowledgeManagementService = Depends(get_knowledge_management_service)):
     """
     Create a new knowledge group.
 
@@ -113,7 +121,7 @@ async def create_group(group: CreateKnowledgeGroupRequest, service: KnowledgeSer
 
 
 @router.get("/knowledge/groups/{group_id}", response_model=KnowledgeGroupResponse)
-async def get_group(group_id: str, service: KnowledgeService = Depends(get_knowledge_service)):
+async def get_group(group_id: str, service: KnowledgeManagementService = Depends(get_knowledge_management_service)):
     """
     Retrieve a knowledge group by its group ID.
 
@@ -141,21 +149,22 @@ async def get_group(group_id: str, service: KnowledgeService = Depends(get_knowl
 
 
 @router.post("/knowledge/groups/{group_id}/ingest", status_code=status.HTTP_202_ACCEPTED)
-async def ingest_group(group_id: str, background_tasks: BackgroundTasks, service: KnowledgeService = Depends(get_knowledge_service)):
+async def ingest_group(group_id: str, background_tasks: BackgroundTasks, km_service: KnowledgeManagementService = Depends(get_knowledge_management_service), ingestion_service: IngestionService = Depends(get_ingestion_service)):
     """
     Initiate the ingestion process for a specific knowledge group.
 
     Args:
         group_id: The ID of the knowledge group to ingest
-        service: Service dependency injection
+        km_service: Knowledge management service dependency injection
+        ingestion_service: Ingestion service dependency injection
     Returns:
         Acknowledgment of ingestion initiation
     """
 
     try:
-        group = await service.find_knowledge_group(group_id)
+        group = await km_service.find_knowledge_group(group_id)
 
-        background_tasks.add_task(service.ingest_knowledge_group, group)
+        background_tasks.add_task(ingestion_service.ingest_knowledge_group, group)
 
         return {"message": f"Ingestion for knowledge group '{group_id}' has been initiated."}
     except KnowledgeGroupNotFoundError as err:
